[Modified] OC-1110  #comment Refactored WS event handlers. Put all necessary values during handshake #time 1h
