+++ src/backend/src/main/java/com/becon/opencelium/backend/configuration/WebSocketConfig.java
@@ -1,7 +1,7 @@
 package com.becon.opencelium.backend.configuration;
 
 import com.becon.opencelium.backend.execution.socket.WebSocketHandshakeInterceptor;
-import com.becon.opencelium.backend.execution.socket.handler.WebSocketEventHandler;
+import com.becon.opencelium.backend.execution.socket.WebSocketEventHandler;
 import com.becon.opencelium.backend.security.JwtTokenUtil;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.Message;
@@ -55,7 +55,7 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
     public void registerStompEndpoints(StompEndpointRegistry registry) {
         registry.addEndpoint(PATH)
                 .setAllowedOriginPatterns("*")
-                .addInterceptors(new WebSocketHandshakeInterceptor(jwtTokenUtil)) // populate attribute [userId]
+                .addInterceptors(new WebSocketHandshakeInterceptor(jwtTokenUtil)) // populate attribute [userId, username, oc-sessionId]
                 .withSockJS();
     }
