--- src/backend/src/main/java/com/becon/opencelium/backend/execution/socket/handler/WebSocketEventHandler.java
@@ -1,8 +1,0 @@
-package com.becon.opencelium.backend.execution.socket.handler;
-
-import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
-
-public interface WebSocketEventHandler {
-    void handleConnect(StompHeaderAccessor accessor);
-    void handleDisconnect(StompHeaderAccessor accessor);
-}
